requirements.txt
fastapi[all]
pyrebase4
pyngrok
ddsp
tensorflow-probability==0.15.0